# svc-transactions/docs/swagger.yaml
basePath: /v1
definitions:
  transactions.DepositRequest:
    properties:
      amount:
        type: integer
      phone_number:
        description: ReferenceID string `json:"reference_id"`
        type: string
      reference_id:
        type: string
    type: object
  transactions.DepositResponse:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      wallet_id:
        type: string
    type: object
  transactions.TransferRequest:
    properties:
      amount:
        type: integer
      receiver_phone:
        type: string
      reference_id:
        type: string
      sender_phone:
        description: "ReferenceID \tstring \t`json:\"reference_id\"`"
        type: string
    type: object
  transactions.TransferResponse:
    properties:
      sender_balance:
        type: integer
      sender_balance_before:
        type: integer
      sender_wallet_id:
        type: string
      status:
        type: string
      transaction_id:
        type: string
    type: object
  transactions.WithdrawalRequest:
    properties:
      amount:
        type: integer
      phone_number:
        description: ReferenceID string `json:"reference_id"`
        type: string
      reference_id:
        type: string
    type: object
  transactions.WithdrawalResponse:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      wallet_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This microservice handles financial transactions.
  title: Transactions Service API
  version: "1.0"
paths:
  /transactions/deposit:
    post:
      consumes:
      - application/json
      description: Validates the incoming payload, checks wallet constraints, and
        credits the wallet balance.
      parameters:
      - description: Deposit Request Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transactions.DepositRequest'
      - description: Unique key to prevent duplicate balance modifications
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Deposit transaction successful & created
          schema:
            $ref: '#/definitions/transactions.DepositResponse'
        "400":
          description: Bad Request (Invalid payload, insufficient balance, capacity
            limit, or phone number)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Wallet Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process a deposit transaction
      tags:
      - Transactions
  /transactions/transfer:
    post:
      consumes:
      - application/json
      description: Validates the incoming payload, checks sender balance, and atomically
        transfers funds from sender to receiver.
      parameters:
      - description: Transfer Request Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transactions.TransferRequest'
      - description: Unique key to prevent duplicate transfers
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Transfer transaction successful & created
          schema:
            $ref: '#/definitions/transactions.TransferResponse'
        "400":
          description: Bad Request (Invalid payload, insufficient balance, or phone
            number)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Sender or Receiver Wallet Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity (Transfer failed due to receiver max capacity)
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests (Concurrency throttle)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process a transfer transaction
      tags:
      - Transactions
  /transactions/withdraw:
    post:
      consumes:
      - application/json
      description: Validates the incoming payload, checks wallet balance, and debits
        the wallet balance.
      parameters:
      - description: Withdrawal Request Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transactions.WithdrawalRequest'
      - description: Unique key to prevent duplicate balance modifications
        in: header
        name: Idempotency-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: withdrawal transaction successful & created
          schema:
            $ref: '#/definitions/transactions.WithdrawalResponse'
        "400":
          description: Bad Request (Invalid payload, insufficient balance, or phone
            number)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Wallet Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process a withdrawal transaction
      tags:
      - Transactions
swagger: "2.0"